cmake: Fix build errors and warnings
The rework addresses multiple build
errors and warnings encountered under
strict compilation flags (e.g., -Werror).

The following issues were resolved:
- Pointer type errors in gzflush calls
are now wrapped in #ifdef DLT_LOGSTORAGE_USE_GZIP
to ensure correct compilation when GZIP support is enabled.
- String truncation and buffer overflow warnings
in logstorage path and filename handling are
fixed by using snprintf and proper length checks.
- The inline keyword was removed from the declaration
of dlt_user_is_logLevel_enabled to match its definition
and resolve unused inline warnings.
- The node_id buffer in passive node control is now
properly null-terminated after strncpy to prevent
truncation warnings.
- The argument order for calloc in dlt-example-user.c
was corrected to avoid transposed-args warnings.

+++ include/dlt/dlt_client.h
@@ -88,16 +88,16 @@ typedef enum
 
 typedef struct
 {
-    DltReceiver receiver;  /**< receiver pointer to dlt receiver structure */
-    int sock;              /**< sock Connection handle/socket */
-    char *servIP;          /**< servIP IP adress/Hostname of interface */
-    char *hostip;          /**< hostip IP address of UDP host receiver interface */
-    int port;              /**< Port for TCP connections (optional) */
-    char *serialDevice;    /**< serialDevice Devicename of serial device */
-    char *socketPath;      /**< socketPath Unix socket path */
-    char ecuid[4];         /**< ECUiD */
-    speed_t baudrate;      /**< baudrate Baudrate of serial interface, as speed_t */
-    DltClientMode mode;    /**< mode DltClientMode */
+    DltReceiver receiver;      /**< receiver pointer to dlt receiver structure */
+    int sock;                  /**< sock Connection handle/socket */
+    char *servIP;              /**< servIP IP adress/Hostname of interface */
+    char *hostip;              /**< hostip IP address of UDP host receiver interface */
+    uint16_t  port;            /**< Port for TCP connections (optional) */
+    char *serialDevice;        /**< serialDevice Devicename of serial device */
+    char *socketPath;          /**< socketPath Unix socket path */
+    char ecuid[4];             /**< ECUiD */
+    speed_t baudrate;          /**< baudrate Baudrate of serial interface, as speed_t */
+    DltClientMode mode;        /**< mode DltClientMode */
     int send_serial_header;    /**< (Boolean) Send DLT messages with serial header */
     int resync_serial_header;  /**< (Boolean) Resync to serial header on all connection */
 } DltClient;
