+++ src/daemon/dlt_daemon_serial.c
@@ -83,13 +83,13 @@ int dlt_daemon_serial_send(int sock,
     /* Send data */
 
     if (data1 && (size1 > 0)) {
-        if (0 > write(sock, data1, size1)) {
+        if (0 > write(sock, data1, (size_t)size1)) {
             return DLT_DAEMON_ERROR_SEND_FAILED;
         }
     }
 
     if (data2 && (size2 > 0)) {
-        if (0 > write(sock, data2, size2)) {
+        if (0 > write(sock, data2, (size_t)size2)) {
             return DLT_DAEMON_ERROR_SEND_FAILED;
         }
     }
